Convert the existing commented out self-test for openmpsort into a unit test
